Image Upload, createProfile API, Create User DB
Added an Image Uploader with multer.

Adjusted the express to recognize the url enconded and json

Added a method for createProfile in the API which call the correct method with the params set.

Now it's creating the user in the db, saving the picture in the backend and redirectioning to the dashboard.

Added a ignore in the watch because it was restarting before it could save the user in the db